standardize use of isHoursActive and fix double rendering stats
Previously Statistics and HoursDisplay created similiar streams to know
which (hours or days) was active; this uses one stream that both share.

Also, previously there was an (edge case) issue with rendering the
Statistics component. Because the model was combine of (at least)
whichHour, whichDay, and isHoursActive (which was a merge of whichHour
and whichDay, mapped to true and false respectively), changing the hour
or day would result in two stream events (the whichHour/whichDay, and
the isHoursActive). If the isHoursActive stream was listed last, going
from day 2 at 1am to day 0 (current) would result in the statistics
model function receiving receiving first of two events, the correct day
(0) but with isHoursActive still true, and so erroring if the current
hour is passed the hour still remembered (1am). If the isHoursActive was
listed first, going from day 2 at 1am to day 0 (current) would avoid
this problem, but then selecting an hour, the model function would
receive (first of two events) isHourActive as true, but the hour as
still 1am.

Using sampleCombine allows for the combine whichHour and whichDay stream
to be the master, and isActiveHour only emit an event when it emits.

(Alternatively, would it make more sense to not track isHourActive and
just nullify the hour. I would still need sampleCombine in some
capacity, otherwise I would still have the two event misinformation
depending on whether i nullify the hour or set the day first, though
nullifying the hour first would just result in two renders, and not an
error.)