don't update history for no zip
This issue is due to allowing zip to be missing, but filling in the url
with it after the fact.  In the future, this could be worked around
(perhaps with a better ux) by offering dynamic keys for both the user
and app to use (when not present, offer the dynamic key equivalent).
Eg, `/here/today/now/auto` that resolves to
`${autoIp}/${new Date().getDate()}/${new Date.getHour() + 1}/english`